fix: added usePagination hook
